simplify strings lib
